refactored MediaLibraryService, created MediaPlaylistService
