+++ src/services/index.ts
@@ -4,6 +4,7 @@ import MediaCollectionService from './media-collection.service';
 import MediaLibraryService from './media-library.service';
 import MediaLikedTrackService from './media-liked-track.service';
 import MediaPlayerService from './media-player.service';
+import MediaPlaylistService from './media-playlist.service';
 import MediaProviderService from './media-provider.service';
 import NotificationService from './notification.service';
 
@@ -14,6 +15,7 @@ export {
   MediaLibraryService,
   MediaLikedTrackService,
   MediaPlayerService,
+  MediaPlaylistService,
   MediaProviderService,
   NotificationService,
 };
