+++ src/services/media-library.service.ts
@@ -3,20 +3,17 @@ import { Semaphore } from 'async-mutex';
 
 import { AppEnums, MediaLibraryActions, MediaTrackCoverPictureImageDataType } from '../enums';
 import store from '../store';
-import { AppError, DataStoreInputData, DataStoreUpdateData } from '../types';
-import { DatastoreUtils, MediaUtils } from '../utils';
+import { DataStoreInputData } from '../types';
+import { MediaUtils } from '../utils';
 
 import AppService from './app.service';
 import MediaPlayerService from './media-player.service';
-import I18nService from './i18n.service';
-import NotificationService from './notification.service';
 
 import {
   MediaAlbumDatastore,
   MediaArtistDatastore,
   MediaProviderDatastore,
   MediaTrackDatastore,
-  MediaPlaylistDatastore,
 } from '../datastores';
 
 import {
@@ -25,50 +22,18 @@ import {
   IMediaArtist,
   IMediaArtistData,
   IMediaPicture,
-  IMediaPlaylist,
-  IMediaPlaylistData,
-  IMediaPlaylistInputData,
-  IMediaPlaylistTrack,
-  IMediaPlaylistTrackData,
-  IMediaPlaylistTrackInputData,
-  IMediaPlaylistTrackUpdateData,
-  IMediaPlaylistUpdateData,
   IMediaTrack,
   IMediaTrackData,
 } from '../interfaces';
 
 export type MediaSyncFunction = () => Promise<void>;
 
-export type MediaSearchResults = {
-  tracks: IMediaTrack[],
-  artists: IMediaArtist[],
-  albums: IMediaAlbum[],
-  playlists: IMediaPlaylist[],
-};
-
 const debug = require('debug')('app:service:media_library_service');
-
-export class MediaLibraryPlaylistDuplicateTracksError extends AppError {
-  existingTrackDataList: IMediaPlaylistTrackInputData[] = [];
-  newTrackDataList: IMediaPlaylistTrackInputData[] = [];
-
-  constructor(
-    existingTrackDataList: IMediaPlaylistTrackInputData[],
-    newTrackDataList: IMediaPlaylistTrackInputData[] = [],
-  ) {
-    super('Duplicate tracks found in playlist');
-    this.name = 'MediaLibraryPlaylistDuplicateTracksError';
-    this.existingTrackDataList = existingTrackDataList;
-    this.newTrackDataList = newTrackDataList;
-  }
-}
 
 class MediaLibraryService {
   readonly mediaPictureScaleWidth = 500;
   readonly mediaPictureScaleHeight = 500;
   private readonly mediaSyncLock = new Semaphore(1);
-
-  // sync API
 
   async syncMedia(mediaProviderIdentifier: string, syncFn: MediaSyncFunction): Promise<void> {
     await this.mediaSyncLock.runExclusive(async () => {
@@ -188,17 +153,6 @@ class MediaLibraryService {
     return this.buildMediaTrack(mediaTrackData, true);
   }
 
-  // search API
-
-  async search(query: string): Promise<MediaSearchResults> {
-    return {
-      tracks: await this.searchTracksByName(query),
-      albums: await this.searchAlbumsByName(query),
-      artists: await this.searchArtistsByName(query),
-      playlists: await this.searchPlaylistsByName(query),
-    };
-  }
-
   async searchTracksByName(query: string): Promise<IMediaTrack[]> {
     const tracks = await MediaTrackDatastore.findMediaTracks({
       track_name: {
@@ -229,18 +183,6 @@ class MediaLibraryService {
     return this.buildMediaArtists(artists);
   }
 
-  async searchPlaylistsByName(query: string): Promise<IMediaPlaylist[]> {
-    const playlists = await MediaPlaylistDatastore.findMediaPlaylists({
-      name: {
-        $regex: new RegExp(query, 'i'),
-      },
-    });
-
-    return this.buildMediaPlaylists(playlists);
-  }
-
-  // fetch API
-
   async getMediaTrack(mediaTrackId: string): Promise<IMediaTrack | undefined> {
     const mediaTrackData = await MediaTrackDatastore.findMediaTrack({
       id: mediaTrackId,
@@ -256,23 +198,6 @@ class MediaLibraryService {
     });
 
     return mediaTrackData ? this.buildMediaTrack(mediaTrackData) : undefined;
-  }
-
-  async getMediaPlaylist(mediaPlaylistId: string): Promise<IMediaPlaylist | undefined> {
-    const mediaPlaylistData = await MediaPlaylistDatastore.findMediaPlaylist({
-      id: mediaPlaylistId,
-    });
-
-    return mediaPlaylistData ? this.buildMediaPlaylist(mediaPlaylistData) : undefined;
-  }
-
-  async getMediaPlaylistTracks(mediaPlaylistId: string): Promise<IMediaPlaylistTrack[]> {
-    const playlist = await this.getMediaPlaylist(mediaPlaylistId);
-    if (!playlist) {
-      throw new Error(`MediaLibraryService encountered error at getMediaPlaylistTracks - Playlist not found - ${mediaPlaylistId}`);
-    }
-
-    return this.buildMediaPlaylistTracks(playlist.tracks);
   }
 
   async getMediaAlbumTracks(mediaAlbumId: string): Promise<IMediaTrack[]> {
@@ -315,82 +240,6 @@ class MediaLibraryService {
     const mediaAlbums = await Promise.all(mediaAlbumDataList.map(mediaAlbumData => this.buildMediaAlbum(mediaAlbumData)));
     return MediaUtils.sortMediaAlbums(mediaAlbums);
   }
-
-  async getMediaPlaylists(): Promise<IMediaPlaylist[]> {
-    const mediaPlaylistsDataList = await MediaPlaylistDatastore.findMediaPlaylists();
-
-    const mediaPlaylists = await Promise.all(
-      mediaPlaylistsDataList.map(mediaPlaylistData => this.buildMediaPlaylist(mediaPlaylistData)),
-    );
-
-    return MediaUtils.sortMediaPlaylists(mediaPlaylists);
-  }
-
-  // create API
-
-  async createMediaPlaylist(mediaPlaylistInputData?: IMediaPlaylistInputData): Promise<IMediaPlaylist> {
-    const inputData: DataStoreInputData<IMediaPlaylistData> = _.defaults(mediaPlaylistInputData, {
-      name: await this.getDefaultNewPlaylistName(),
-      tracks: [],
-      created_at: Date.now(),
-    });
-    inputData.tracks = inputData.tracks.map(trackInputData => this.buildMediaPlaylistTrackFromInput(trackInputData));
-
-    const mediaPlaylistData = await MediaPlaylistDatastore.insertMediaPlaylist(inputData);
-    const mediaPlaylist = await this.buildMediaPlaylist(mediaPlaylistData);
-
-    store.dispatch({
-      type: MediaLibraryActions.AddPlaylist,
-      data: {
-        mediaPlaylist,
-      },
-    });
-
-    return mediaPlaylist;
-  }
-
-  /**
-   * @throws MediaLibraryPlaylistDuplicateTracksError
-   */
-  async addMediaPlaylistTracks(mediaPlaylistId: string, mediaPlaylistTrackInputDataList: IMediaPlaylistTrackInputData[], options?: {
-    ignoreExisting?: boolean; // only add new ones
-  }): Promise<IMediaPlaylist> {
-    const {
-      existingInputDataList,
-      newInputDataList,
-    } = await this.getExistingMediaPlaylistTrackInputData(
-      mediaPlaylistId,
-      mediaPlaylistTrackInputDataList,
-    );
-
-    if (!_.isEmpty(existingInputDataList) && !options?.ignoreExisting) {
-      // not allowed to add duplicate tracks, throw error
-      throw new MediaLibraryPlaylistDuplicateTracksError(existingInputDataList, newInputDataList);
-    }
-
-    // all good
-    const mediaPlaylistData = await MediaPlaylistDatastore.addMediaPlaylistTracks(
-      mediaPlaylistId,
-      mediaPlaylistTrackInputDataList.map(trackInputData => this.buildMediaPlaylistTrackFromInput(trackInputData)),
-    );
-
-    const mediaPlaylistUpdated = await this.buildMediaPlaylist(mediaPlaylistData);
-
-    store.dispatch({
-      type: MediaLibraryActions.AddPlaylist,
-      data: {
-        mediaPlaylist: mediaPlaylistUpdated,
-      },
-    });
-
-    NotificationService.showMessage(I18nService.getString('message_added_to_playlist', {
-      playlistName: mediaPlaylistUpdated.name,
-    }));
-
-    return mediaPlaylistUpdated;
-  }
-
-  // load API
 
   loadMediaAlbums(): void {
     this
@@ -445,81 +294,6 @@ class MediaLibraryService {
         });
       });
   }
-
-  loadMediaPlaylists(): void {
-    this
-      .getMediaPlaylists()
-      .then((mediaPlaylists) => {
-        store.dispatch({
-          type: MediaLibraryActions.SetPlaylists,
-          data: {
-            mediaPlaylists,
-          },
-        });
-      });
-  }
-
-  loadMediaPlaylist(mediaPlaylistId: string): void {
-    this
-      .getMediaPlaylist(mediaPlaylistId)
-      .then((mediaPlaylist) => {
-        store.dispatch({
-          type: MediaLibraryActions.SetPlaylist,
-          data: {
-            mediaPlaylist,
-          },
-        });
-      });
-  }
-
-  // update API
-
-  async updateMediaPlaylist(mediaPlaylistId: string, mediaPlaylistUpdateData: IMediaPlaylistUpdateData): Promise<IMediaPlaylist> {
-    const mediaPlaylistData = await MediaPlaylistDatastore.updateMediaPlaylist(mediaPlaylistId, await this.buildMediaPlaylistUpdateDataFromInput(mediaPlaylistId, mediaPlaylistUpdateData));
-    const mediaPlaylist = await this.buildMediaPlaylist(mediaPlaylistData);
-
-    store.dispatch({
-      type: MediaLibraryActions.AddPlaylist,
-      data: {
-        mediaPlaylist,
-      },
-    });
-
-    return mediaPlaylist;
-  }
-
-  // delete API
-
-  async deleteMediaPlaylist(mediaPlaylistId: string): Promise<void> {
-    await MediaPlaylistDatastore.deleteMediaPlaylist({
-      id: mediaPlaylistId,
-    });
-
-    store.dispatch({
-      type: MediaLibraryActions.RemovePlaylist,
-      data: {
-        mediaPlaylistId,
-      },
-    });
-
-    NotificationService.showMessage(I18nService.getString('message_playlist_deleted'));
-  }
-
-  async deleteMediaPlaylistTracks(mediaPlaylistId: string, mediaPlaylistTrackIds: string[]): Promise<IMediaPlaylist> {
-    const mediaPlaylistData = await MediaPlaylistDatastore.deleteMediaPlaylistTracks(mediaPlaylistId, mediaPlaylistTrackIds);
-    const mediaPlaylist = await this.buildMediaPlaylist(mediaPlaylistData);
-
-    store.dispatch({
-      type: MediaLibraryActions.AddPlaylist,
-      data: {
-        mediaPlaylist,
-      },
-    });
-
-    return mediaPlaylist;
-  }
-
-  // private API
 
   private async startMediaTrackSync(mediaProviderIdentifier: string): Promise<void> {
     const mediaProviderData = await MediaProviderDatastore.findMediaProviderByIdentifier(mediaProviderIdentifier);
@@ -708,123 +482,6 @@ class MediaLibraryService {
         $lt: mediaSyncStartTimestamp,
       },
     });
-  }
-
-  private async buildMediaPlaylist(mediaPlaylistData: IMediaPlaylistData) {
-    return _.assign(mediaPlaylistData, {});
-  }
-
-  private async buildMediaPlaylists(mediaPlaylistDataList: IMediaPlaylistData[]) {
-    return Promise.all(mediaPlaylistDataList.map((mediaPlaylistData: any) => this.buildMediaPlaylist(mediaPlaylistData)));
-  }
-
-  private async buildMediaPlaylistTracks(mediaPlaylistTrackDataList: IMediaPlaylistTrackData[]): Promise<IMediaPlaylistTrack[]> {
-    // TODO: Added a hack here to ignore playlist tracks that could not be found locally anymore
-    //  Ideally, such entries should have a status like "unavailable"
-
-    // @ts-ignore
-    return Promise
-      .all(mediaPlaylistTrackDataList.map(mediaPlaylistTrackData => this.buildMediaPlaylistTrack(mediaPlaylistTrackData)))
-      .then(mediaPlaylistTracks => mediaPlaylistTracks.filter(mediaPlaylistTrack => !_.isNil(mediaPlaylistTrack)));
-  }
-
-  private async buildMediaPlaylistTrack(mediaPlaylistTrackData: IMediaPlaylistTrackData): Promise<IMediaPlaylistTrack | undefined> {
-    const mediaTrack = await this.getMediaTrackForProvider(mediaPlaylistTrackData.provider, mediaPlaylistTrackData.provider_id);
-    if (!mediaTrack) {
-      return undefined;
-    }
-
-    return _.assign({}, mediaPlaylistTrackData, mediaTrack);
-  }
-
-  private async getDefaultNewPlaylistName(): Promise<string> {
-    const mediaPlaylistsCount = await MediaPlaylistDatastore.countMediaPlaylists();
-
-    return `${I18nService.getString('label_new_playlist_default_name', {
-      playlistCount: (mediaPlaylistsCount + 1).toString(),
-    })}`;
-  }
-
-  private buildMediaPlaylistTrackFromInput(trackInputData: IMediaPlaylistTrackInputData): IMediaPlaylistTrackData {
-    return {
-      playlist_track_id: DatastoreUtils.generateId(),
-      provider: trackInputData.provider,
-      provider_id: trackInputData.provider_id,
-      added_at: Date.now(),
-    };
-  }
-
-  private async buildMediaPlaylistUpdateDataFromInput(playlistId: string, playlistUpdateData: IMediaPlaylistUpdateData): Promise<DataStoreUpdateData<IMediaPlaylistData>> {
-    const data: DataStoreUpdateData<IMediaPlaylistData> = {};
-    if (playlistUpdateData.name) {
-      data.name = playlistUpdateData.name;
-    }
-    if (playlistUpdateData.cover_picture) {
-      data.cover_picture = playlistUpdateData.cover_picture;
-    }
-    if (playlistUpdateData.tracks) {
-      data.tracks = await this.buildMediaPlaylistTrackUpdateDataFromInput(playlistId, playlistUpdateData.tracks);
-    }
-
-    return data;
-  }
-
-  private async buildMediaPlaylistTrackUpdateDataFromInput(playlistId: string, playlistTrackUpdateDataList: IMediaPlaylistTrackUpdateData[]): Promise<IMediaPlaylistTrackData[]> {
-    // we got tracks to update, we only get playlist_track_id in the order we required
-    // we also can have deleted ids, no addition is allowed
-    // build the new set of playlist tracks in order we require and set them directly
-    const playlistData = await MediaPlaylistDatastore.findMediaPlaylist({
-      id: playlistId,
-    });
-    if (!playlistData) {
-      throw new Error(`MediaLibraryService encountered error at buildMediaPlaylistTrackUpdateDataFromInput - Could not find playlist - ${playlistId}`);
-    }
-
-    const playlistUpdatedTracks: IMediaPlaylistTrackData[] = [];
-    playlistTrackUpdateDataList.forEach((trackUpdateData) => {
-      const playlistTrackData = playlistData.tracks.find(
-        trackData => trackData.playlist_track_id === trackUpdateData.playlist_track_id,
-      );
-      // no addition allowed
-      if (!playlistTrackData) {
-        throw new Error(`MediaLibraryService encountered error at buildMediaPlaylistTrackUpdateDataFromInput - Could not find track in playlist - ${trackUpdateData.playlist_track_id}`);
-      }
-
-      playlistUpdatedTracks.push(playlistTrackData);
-    });
-
-    return playlistUpdatedTracks;
-  }
-
-  private async getExistingMediaPlaylistTrackInputData(mediaPlaylistId: string, mediaPlaylistTrackInputDataList: IMediaPlaylistTrackInputData[]): Promise<{
-    existingInputDataList: IMediaPlaylistTrackInputData[],
-    newInputDataList: IMediaPlaylistTrackInputData[],
-  }> {
-    const playlist = await this.getMediaPlaylist(mediaPlaylistId);
-    if (!playlist) {
-      throw new Error(`MediaLibraryService encountered error at getExistingMediaPlaylistTrackInputData - Playlist not found - ${mediaPlaylistId}`);
-    }
-
-    const existingInputDataList: IMediaPlaylistTrackInputData[] = [];
-    const newInputDataList: IMediaPlaylistTrackInputData[] = [];
-
-    mediaPlaylistTrackInputDataList.forEach((trackData) => {
-      const playlistTrack = playlist.tracks.find(
-        data => data.provider === trackData.provider && data.provider_id === trackData.provider_id,
-      );
-      if (playlistTrack) {
-        // existing track
-        existingInputDataList.push(trackData);
-      } else {
-        // new track
-        newInputDataList.push(trackData);
-      }
-    });
-
-    return {
-      existingInputDataList,
-      newInputDataList,
-    };
   }
 }
